requirements.txt
assistant-fulfillment-helper==1.0.3
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-WTF==1.0.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
mysql-connector-python==8.0.31
packaging==22.0
protobuf==3.20.1
PyMySQL==1.0.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.45
webargs==8.2.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
